+++ python/boomer/common/rules.pxd
@@ -30,7 +30,7 @@ cdef class Body:
 
     cdef bint covers(self, float32[::1] example)
 
-    cdef bint covers_sparse(self, float32[::1] example_data, intp[::1] example_indices, float32[::1] tmp_array1,
+    cdef bint covers_sparse(self, float32[::1] example_data, uint32[::1] example_indices, float32[::1] tmp_array1,
                             uint32[::1] tmp_array2, uint32 n)
 
 
@@ -40,7 +40,7 @@ cdef class EmptyBody(Body):
 
     cdef bint covers(self, float32[::1] example)
 
-    cdef bint covers_sparse(self, float32[::1] example_data, intp[::1] example_indices, float32[::1] tmp_array1,
+    cdef bint covers_sparse(self, float32[::1] example_data, uint32[::1] example_indices, float32[::1] tmp_array1,
                             uint32[::1] tmp_array2, uint32 n)
 
 
@@ -68,7 +68,7 @@ cdef class ConjunctiveBody(Body):
 
     cdef bint covers(self, float32[::1] example)
 
-    cdef bint covers_sparse(self, float32[::1] example_data, intp[::1] example_indices, float32[::1] tmp_array1,
+    cdef bint covers_sparse(self, float32[::1] example_data, uint32[::1] example_indices, float32[::1] tmp_array1,
                             uint32[::1] tmp_array2, uint32 n)
 
 
@@ -115,7 +115,7 @@ cdef class Rule:
 
     cdef predict(self, float32[:, ::1] x, float64[:, ::1] predictions, uint8[:, ::1] mask=*)
 
-    cdef predict_csr(self, float32[::1] x_data, intp[::1] x_row_indices, intp[::1] x_col_indices, intp num_features,
+    cdef predict_csr(self, float32[::1] x_data, uint32[::1] x_row_indices, uint32[::1] x_col_indices, intp num_features,
                      float32[::1] tmp_array1, uint32[::1] tmp_array2, uint32 n, float64[:, ::1] predictions,
                      uint8[:, ::1] mask=*)
 
@@ -128,7 +128,7 @@ cdef class RuleModel:
 
     cdef float64[:, ::1] predict(self, float32[:, ::1] x, intp num_labels)
 
-    cdef float64[:, ::1] predict_csr(self, float32[::1] x_data, intp[::1] x_row_indices, intp[::1] x_col_indices,
+    cdef float64[:, ::1] predict_csr(self, float32[::1] x_data, uint32[::1] x_row_indices, uint32[::1] x_col_indices,
                                      intp num_features, intp num_labels)
 
 
@@ -146,7 +146,7 @@ cdef class RuleList(RuleModel):
 
     cdef float64[:, ::1] predict(self, float32[:, ::1] x, intp num_labels)
 
-    cdef float64[:, ::1] predict_csr(self, float32[::1] x_data, intp[::1] x_row_indices, intp[::1] x_col_indices,
+    cdef float64[:, ::1] predict_csr(self, float32[::1] x_data, uint32[::1] x_row_indices, uint32[::1] x_col_indices,
                                      intp num_features, intp num_labels)
